+++ TeslaSolarCharger/Shared/Dtos/IndexRazor/PvValues/DtoPvValues.cs
@@ -2,10 +2,10 @@
 
 public class DtoPvValues
 {
-    public int? InverterPower { get; set; }
-    public int? GridPower { get; set; }
-    public int? HomeBatteryPower { get; set; }
-    public int? HomeBatterySoc { get; set; }
+    public double? InverterPower { get; set; }
+    public double? GridPower { get; set; }
+    public double? HomeBatteryPower { get; set; }
+    public double? HomeBatterySoc { get; set; }
     public int? CarCombinedChargingPowerAtHome { get; set; }
     public DateTime LastUpdated { get; set; }
 }
